Extracted Parquet projection related tests into standalone file
